# services/backend/openapi/openapi.json
{
  "components": {
    "schemas": {
      "AccountPatch": {
        "properties": {
          "displayName": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          }
        },
        "title": "AccountPatch",
        "type": "object"
      },
      "AccountProfile": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "displayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "providers": {
            "items": {
              "type": "string"
            },
            "title": "Providers",
            "type": "array"
          }
        },
        "required": [
          "id",
          "displayName",
          "email",
          "providers",
          "permissions",
          "createdAt"
        ],
        "title": "AccountProfile",
        "type": "object"
      },
...
      "CatalogMeta": {
        "properties": {
          "lastSuccessfulSync": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastsuccessfulsync"
          },
          "productCount": {
            "title": "Productcount",
            "type": "integer"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/CatalogSource"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "productCount",
          "lastSuccessfulSync",
          "sources"
        ],
        "title": "CatalogMeta",
        "type": "object"
      },
      "CatalogSource": {
        "properties": {
          "attribution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribution"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "licenseName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Licensename"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "portalUrl": {
            "title": "Portalurl",
            "type": "string"
          }
        },
        "required": [
          "code",
          "name",
          "portalUrl",
          "licenseName",
          "attribution"
        ],
        "title": "CatalogSource",
        "type": "object"
      },
      "CursorPage_DrugSafetyRule_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DrugSafetyRule"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "title": "CursorPage[DrugSafetyRule]",
        "type": "object"
      },
      "CursorPage_DrugSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DrugSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "title": "CursorPage[DrugSummary]",
        "type": "object"
      },
      "DeleteAccountRequest": {
        "properties": {
          "reauthGrant": {
            "title": "Reauthgrant",
            "type": "string"
          }
        },
        "required": [
          "reauthGrant"
        ],
        "title": "DeleteAccountRequest",
        "type": "object"
      },
...
      "DrugDetail": {
        "properties": {
          "appearance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appearance"
          },
          "efficacy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Efficacy"
          },
          "identification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DrugAppearanceInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "identificationVariants": {
            "items": {
              "$ref": "#/components/schemas/DrugAppearanceInfo"
            },
            "title": "Identificationvariants",
            "type": "array"
          },
          "imageUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imageurl"
          },
          "ingredients": {
            "items": {
              "type": "string"
            },
            "title": "Ingredients",
            "type": "array"
          },
          "interactions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interactions"
          },
          "itemName": {
            "title": "Itemname",
            "type": "string"
          },
          "itemSeq": {
            "title": "Itemseq",
            "type": "string"
          },
          "manufacturer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manufacturer"
          },
          "permitDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permitdate"
          },
          "precautions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Precautions"
          },
          "professionalCategory": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Professionalcategory"
          },
          "safetyOverview": {
            "$ref": "#/components/schemas/DrugSafetyOverview"
          },
          "sideEffects": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sideeffects"
          },
          "sourceUpdatedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceupdatedat"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/DrugSourceAttribution"
            },
            "title": "Sources",
            "type": "array"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "storageMethod": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storagemethod"
          },
          "useMethod": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usemethod"
          },
          "warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warning"
          }
        },
        "required": [
          "itemSeq",
          "itemName",
          "manufacturer",
          "status",
          "professionalCategory",
          "permitDate",
          "storageMethod",
          "appearance",
          "imageUrl",
          "identification",
          "identificationVariants",
          "safetyOverview",
          "ingredients",
          "efficacy",
          "useMethod",
          "warning",
          "precautions",
          "interactions",
          "sideEffects",
          "sourceUpdatedAt",
          "sources"
        ],
        "title": "DrugDetail",
        "type": "object"
      },
...
      "DrugSafetyOverview": {
        "properties": {
          "categories": {
            "items": {
              "$ref": "#/components/schemas/DrugSafetyCategory"
            },
            "title": "Categories",
            "type": "array"
          },
          "totalCount": {
            "title": "Totalcount",
            "type": "integer"
          }
        },
        "required": [
          "totalCount",
          "categories"
        ],
        "title": "DrugSafetyOverview",
        "type": "object"
      },
...
      "DrugSummary": {
        "properties": {
          "itemName": {
            "title": "Itemname",
            "type": "string"
          },
          "itemSeq": {
            "title": "Itemseq",
            "type": "string"
          },
          "manufacturer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manufacturer"
          },
          "professionalCategory": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Professionalcategory"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "required": [
          "itemSeq",
          "itemName",
          "manufacturer",
          "status",
          "professionalCategory"
        ],
        "title": "DrugSummary",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Medical Box API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/catalog/meta": {
      "get": {
        "operationId": "catalog_meta_api_v1_catalog_meta_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogMeta"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Catalog Meta",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/drugs/search": {
      "get": {
        "operationId": "search_drugs_api_v1_drugs_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 120,
              "minLength": 2,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPage_DrugSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Search Drugs",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/drugs/{itemSeq}": {
      "get": {
        "operationId": "get_drug_api_v1_drugs__itemSeq__get",
        "parameters": [
          {
            "in": "path",
            "name": "itemSeq",
            "required": true,
            "schema": {
              "title": "Itemseq",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrugDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Drug",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/drugs/{itemSeq}/dur-rules": {
      "get": {
        "operationId": "get_drug_safety_rules_api_v1_drugs__itemSeq__dur_rules_get",
        "parameters": [
          {
            "in": "path",
            "name": "itemSeq",
            "required": true,
            "schema": {
              "title": "Itemseq",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "ruleType",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ruletype"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPage_DrugSafetyRule_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Drug Safety Rules",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/me": {
      "delete": {
        "operationId": "delete_me_api_v1_me_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Me",
        "tags": [
          "account"
        ]
      },
      "get": {
        "operationId": "get_me_api_v1_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountProfile"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Me",
        "tags": [
          "account"
        ]
      },
      "patch": {
        "operationId": "update_me_api_v1_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Me",
        "tags": [
          "account"
        ]
      }
    },
...
  }
}